실수의 개념 이해
실수가 필요한 상황을 보며 실수의 개념 이해해보기

+++ day03/src/date2/Test01.java
@@ -10,7 +10,8 @@ public class Test01 {
 		 int english = 80;
 		 
 		 //계산
-		 int average = ( korean + english ) / 2;
+		double average = ( korean + english ) / 2.0;
+		//int average = ( korean + english )  2;  ->결과 값은 int
 		 
 		 //출력
 		 System.out.println(average);
